+++ socket_client.py
@@ -5,9 +5,8 @@ import websockets
 import cv2
 import numpy as np
 import time
-from main import gemini
+from gemini import gemini
 from key import TOKEN
-from start import PaddleProcessWords
 
 async def video_receiver(websocket, gemini_):
     print("Client connected")
@@ -34,8 +33,6 @@ async def video_receiver(websocket, gemini_):
 
 async def main(gemin):
     #FILL IN THIS IP ADDRESS 
-    start = time.time()
-
 
     async def handler(websocket):
         await video_receiver(websocket, gemin) # Pass the instance
@@ -44,10 +41,52 @@ async def main(gemin):
         print("WebSocket server started")
         await asyncio.Future()  # run forever
 
-    end = time.time()
+
+
+
+
+
+
+
+
 
 if __name__ == "__main__":
     gemin = gemini(TOKEN)
     asyncio.run(main(gemin))
 
 
+'''
+    # below was unused
+
+
+    def PaddleProcessMath():
+    from paddlex import create_pipeline
+
+    pipeline = create_pipeline(pipeline="formula_recognition")
+
+    output = pipeline.predict(
+        input="./Screenshot1.png",
+        use_layout_detection=True ,
+        use_doc_orientation_classify=False, #we can say yes, but prolly inef
+        use_doc_unwarping=False, # we can do yes, but probably inef
+        layout_threshold=0.5,
+        layout_nms=True,
+        layout_unclip_ratio=1.0,
+        layout_merge_bboxes_mode="large"
+    )
+    for res in output:
+        # res.print()
+        # res.save_to_img(save_path="./layout_output/")
+        res.save_to_json(save_path="./layout_output/")
+
+
+
+
+
+
+
+
+
+
+
+'''
